# src/frontend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VenueAxe.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5280/"
    }
  ],
  "paths": {
...
    "/api/admin/lanes/venue/{venueId}": {
      "get": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaneRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaneRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/lanes/{id}": {
      "get": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLaneRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLaneRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLaneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/lanes/{id}/status": {
      "put": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LaneStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/lanes/{id}/regenerate-pairing": {
      "post": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/lanes/{sourceLaneId}/transfer-to/{targetLaneId}": {
      "post": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "sourceLaneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "targetLaneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/lanes/{id}/toggle-active": {
      "put": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ToggleLaneActiveRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ToggleLaneActiveRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ToggleLaneActiveRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/lanes/{id}/upcoming-reservations": {
      "get": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/lanes/venue/{venueId}/available-for-slot": {
      "get": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "durationMinutes",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 60
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneSlotOptionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneSlotOptionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneSlotOptionDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActiveSessionSummaryDto": {
        "required": [
          "sessionId",
          "sessionTitle",
          "startedAt",
          "expiresAt",
          "minutesRemaining",
          "activeRosterJson",
          "currentGame"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionTitle": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "minutesRemaining": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "activeRosterJson": {
            "type": "string"
          },
          "currentGame": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GameStateSnapshot"
              }
            ]
          }
        }
      },
...
      "BookingDto": {
        "required": [
          "id",
          "venueId",
          "bookingReference",
          "status",
          "guestFirstName",
          "guestLastName",
          "guestEmail",
          "guestPhone",
          "partySize",
          "startTime",
          "endTime",
          "totalAmountCents",
          "paidAmountCents",
          "paymentStatus",
          "assignedLaneNumbers",
          "signedWaiverCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "bookingReference": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "guestFirstName": {
            "type": "string"
          },
          "guestLastName": {
            "type": "string"
          },
          "guestEmail": {
            "type": "string"
          },
          "guestPhone": {
            "type": "string"
          },
          "partySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "paidAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "paymentStatus": {
            "type": "string"
          },
          "assignedLaneNumbers": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          "signedWaiverCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bookingTypeId": {
            "type": [
              "null",
              "string"
            ]
          },
          "discountAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          },
          "appliedDiscountCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "squarePaymentId": {
            "type": [
              "null",
              "string"
            ]
          },
          "balanceDueCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "BookingStatus": {
        "type": "integer"
      },
...
      "CreateLaneRequest": {
        "required": [
          "laneNumber",
          "name"
        ],
        "type": "object",
        "properties": {
          "laneNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "maxThrowers": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 6
          }
        }
      },
...
      "GamePlayer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarColor": {
            "type": "string"
          },
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "throwsTaken": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bullseyesHit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "killsHit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "clutchesHit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "killsRemaining": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "killsCalled": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "streak": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "throwHistory": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        }
      },
      "GameStateSnapshot": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "format": "uuid"
          },
          "gameTypeId": {
            "type": "string"
          },
          "gameName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "currentRound": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalRounds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentPlayerIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GamePlayer"
            }
          },
          "winnerPlayerId": {
            "type": [
              "null",
              "string"
            ]
          },
          "winnerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastThrow": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ThrowRecord"
              }
            ]
          },
          "allThrows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThrowRecord"
            }
          },
          "customDataJson": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LaneDto": {
        "required": [
          "id",
          "venueId",
          "laneNumber",
          "name",
          "maxThrowers",
          "currentStatus",
          "tabletPairingCode",
          "screenPairingCode",
          "lastHeartbeatAt",
          "activeSession"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "laneNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "maxThrowers": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentStatus": {
            "$ref": "#/components/schemas/LaneStatus"
          },
          "tabletPairingCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "screenPairingCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastHeartbeatAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "activeSession": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ActiveSessionSummaryDto"
              }
            ]
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "nextBookingToday": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NextBookingSummaryDto"
              }
            ]
          }
        }
      },
...
      "LaneSlotOptionDto": {
        "required": [
          "laneId",
          "laneNumber",
          "name",
          "isAvailable"
        ],
        "type": "object",
        "properties": {
          "laneId": {
            "type": "string",
            "format": "uuid"
          },
          "laneNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "conflictReason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LaneStatus": {
        "type": "integer"
      },
...
      "MatchStatus": {
        "type": "integer"
      },
      "NextBookingSummaryDto": {
        "required": [
          "bookingId",
          "bookingReference",
          "guestName",
          "startTime",
          "endTime",
          "partySize"
        ],
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "bookingReference": {
            "type": "string"
          },
          "guestName": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "partySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          },
          "paidAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          },
          "paymentStatus": {
            "type": "string",
            "default": "Pending"
          }
        }
      },
...
      "TargetZone": {
        "type": "integer"
      },
...
      "ThrowRecord": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string"
          },
          "playerName": {
            "type": "string"
          },
          "zone": {
            "$ref": "#/components/schemas/TargetZone"
          },
          "pointsAwarded": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "isKillCalled": {
            "type": "boolean"
          },
          "isClutchCalled": {
            "type": "boolean"
          },
          "isBullseye": {
            "type": "boolean"
          },
          "isKillshot": {
            "type": "boolean"
          },
          "thrownAt": {
            "type": "string",
            "format": "date-time"
          },
          "targetCellIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ToggleLaneActiveRequest": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "UpdateLaneRequest": {
        "required": [
          "laneNumber",
          "name",
          "maxThrowers",
          "status"
        ],
        "type": "object",
        "properties": {
          "laneNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "maxThrowers": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/LaneStatus"
          }
        }
      },
...
    }
  },
...
}